Test mode.

+++ beam-core/src/main/java/beam/core/diff/Delete.java
@@ -2,6 +2,7 @@ package beam.core.diff;
 
 import beam.core.BeamUI;
 import beam.lang.Resource;
+import beam.lang.ast.scope.State;
 
 public class Delete extends Change {
 
@@ -27,8 +28,10 @@ public class Delete extends Change {
     }
 
     @Override
-    public void execute() {
-        ((Resource) diffable).delete();
+    public void execute(BeamUI ui, State state) {
+        if (!state.isTest()) {
+            ((Resource) diffable).delete();
+        }
     }
 
 }
